* Added new nand emulation feature:
With this feature it's possible to temp change the region of your Wii console. This makes it possible to play those one region only games that wouldn't play on other region consoles. 

Some games that work now are for example:

The Amazing Race (didn't play on PAL consoles)
Hollywood Squares (didn't play on PAL consoles)
Arthur and the Revenge of Maltazard (didn't play on NTSC consoles)

The feature is now working for Wii games if FULL nand emulation is enabled. The Regionswitch only emulation mode is written for and will only work with a yet unreleased feature of davebaol's d2x cIOS

* Updated english.ini
* Updated dutch.ini

Other languages have to change also again :P  




git-svn-id: http://open-wiiflow-mod.googlecode.com/svn/trunk@206 5d90fa2f-112f-2c21-deeb-19951a5ba7ed