Added an Age Lock to the parental control menu.
You can select an age from 2-19, where any game rated above is not displayed.  It uses the age from the region of the game.  19 is all games. 2 is basically none, because the rating systems start at 3.  This operates independently from the current per game lock, so you can still use both.  Meaning both conditions have to allow the game, before it will display.

You can override the rating by creating an age_lock.ini file.  Just add the id under the domain heading and set it to an age.  To set Call of Duty: Black Ops to an age of 5 you would enter this:
[GAMES]
SC7E52=5

There is also an age_lock_default in the wiiflow.ini file under [GENERAL].  This will allow you to specify a default rating for games with no rating info.


git-svn-id: http://open-wiiflow-mod.googlecode.com/svn/trunk@351 5d90fa2f-112f-2c21-deeb-19951a5ba7ed